# oapi-spec.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Spolyr
servers:
  - url: http://127.0.0.1:8080/api
paths:
...
  /auth/refresh:
    get:
      tags:
        - auth
      summary: Refresh JWT access token
      security:
        - cookieAuth: []
      responses:
        201:
          description: OK
        401:
          description: No refresh token provided
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt